mv relevant functions to common module for reuse when doing transcription for screen recordings
