+++ tiktok_reporter_analysis/classify_reported.py
@@ -1,44 +1,11 @@
 import argparse
-import os
 from tempfile import NamedTemporaryFile
 
-import numpy as np
-import pandas as pd
-import whisper
-from moviepy.editor import VideoFileClip
-from PIL import Image
-
-from tiktok_reporter_analysis.common import extract_frames, multi_modal_analysis
-
-
-def extract_frames_n_audio(video_path, audio_path, results_path, num_frames=2):
-    clip = VideoFileClip(video_path)
-    audio = clip.audio
-    audio.write_audiofile(audio_path)
-
-    n_frames_in_video = int(clip.fps * clip.duration)
-    frame_timestamps = np.linspace(0, clip.duration, n_frames_in_video)
-    selected_frames = extract_frames(frame_timestamps)
-
-    frames = {
-        np.where(frame_timestamps == time)[0][0]: Image.fromarray(clip.get_frame(time)) for time in selected_frames
-    }
-
-    # Create a pandas dataframe with frame number and timestamp
-    df = pd.DataFrame({"frame": list(frames.keys()), "timestamp": selected_frames})
-    df["timestamp"] = df["timestamp"] * 1000  # Convert to milliseconds
-
-    # Write the dataframe to a csv file
-    df.to_csv(os.path.join(results_path, "frames_n_timestamps.csv"), index=False)
-
-    return frames
-
-
-def extract_transcript(audio_path):
-    whisper_model = whisper.load_model("base")
-    print(whisper_model.device)
-    transcript = whisper_model.transcribe(audio_path)
-    return transcript
+from tiktok_reporter_analysis.common import (
+    extract_frames_n_audio,
+    extract_transcript,
+    multi_modal_analysis,
+)
 
 
 def classify_reported(video_path, results_path, testing=False):
